generated_files/file_20250302_002056_174h4u.py
"""
Auto-generated Python file for GitHub contributions.
Generated at: 2025-03-02 00:21:00

This file contains Python code for data processing with error handling.
"""

from typing import Dict, List, Any, Union
import json
from datetime import datetime



import pandas as pd
from typing import Dict, Union

def process_data(file_path: str) -> Union[Dict[str, Dict[str, float]], str]:
    """
    Process a CSV file to calculate the mean and standard deviation for each column.

    Parameters:
    file_path (str): The path to the CSV file.

    Returns:
    Union[Dict[str, Dict[str, float]], str]: A dictionary containing the mean and standard deviation for each column,
    or an error message if an error occurs.

    Raises:
    FileNotFoundError: If the file does not exist.
    pd.errors.EmptyDataError: If the file is empty.
    pd.errors.ParserError: If there is an error parsing the file.
    """
    try:
        # Read the CSV file
        data = pd.read_csv(file_path)

        # Initialize a dictionary to store the results
        results = {}

        # Iterate through each column and calculate mean and standard deviation
        for column in data.columns:
            mean = data[column].mean()
            std = data[column].std()
            results[column] = {'mean': mean, 'std': std}

        return results

    except FileNotFoundError:
        return "Error: The file was not found."
    except pd.errors.EmptyDataError:
        return "Error: The file is empty."
    except pd.errors.ParserError:
        return "Error: There was an error parsing the file."
    except Exception as e:
        return f"An unexpected error occurred: {str(e)}"

# Example usage:
# result = process_data('data.csv')
# print(result)

if __name__ == "__main__":
    test_cases = None
    # Check which function exists in the generated code
    if "process_data" in generated_code:
        test_cases = [
            {"name": "Test User", "age": 25, "status": "ACTIVE"},
            {"input": "Hello", "value": 123, "type": "test"}
        ]
        try:
            for data in test_cases:
                result = process_data(data)
                print(f"\nTest input:  {json.dumps(data, indent=2)}")
                print(f"Test output: {json.dumps(result, indent=2)}")
        except Exception as e:
            print(f"Error: {str(e)}")
    elif "analyze_text" in generated_code:
        test_cases = [
            "Hello World! This is a Test String.",
            "UPPER lower 12345 !@#$%"
        ]
        try:
            for text in test_cases:
                result = analyze_text(text)
                print(f"\nTest input:  {json.dumps(text)}")
                print(f"Test output: {json.dumps(result, indent=2)}")
        except Exception as e:
            print(f"Error: {str(e)}")
    elif "transform_list" in generated_code:
        test_cases = [
            ["item1", "ITEM2", "Item3"],
            ["Python", "JAVA", "TypeScript"]
        ]
        try:
            for items in test_cases:
                result = transform_list(items)
                print(f"\nTest input:  {json.dumps(items)}")
                print(f"Test output: {json.dumps(result, indent=2)}")
        except Exception as e:
            print(f"Error: {str(e)}")
