requirements.txt
pytest
requests
flask
python-dotenv
openai
